### 2025-04-05 #7680 @ekourlit
Glad to edit this paper @kyleniemeyer -- I have quite of a background with Geant4 in LHC.
