### 2025-04-07 #7680 @niess
Hello @ekourlit,

Thank you for taking care of this submission.

Concerning potential reviewers, I am not quite convinced by the automatic suggestions.

I would say that most people having both a background in High Energy Physics (HEP) and computing would match. But, looking at previous JOSS submissions (with "hep" as keyword) there are only two relevant matches, one of which (Go-HEP) originates from my lab (LPC), thus being a Conflict Of Interest (COI) as far as I understand.

A similar project to this submission is [opengate](https://github.com/OpenGATE/opengate), but there might be COIs as well, since some contributors are from my lab, or institute (IN2P3).

Otherwise, looking at GitHub/Geant4 related projects/contributors, [Ben Morgan](https://github.com/drbenmorgan) or [Christopher Poole](https://github.com/christopherpoole) would be great fits as reviewers, to me, if I could propose "external" people? (and if they would be willing to do so). In would not see COIs in these cases.
