### 2025-04-15 2414764
Add pid and energy to detailed deposit
