### 2025-06-10 b7a00ba
Add tid to detailed deposit
