+++ src/simulation.rs
@@ -399,6 +399,7 @@ impl<'a> RunAgent<'a> {
     pub fn push_deposit(
         &mut self,
         volume: *const ffi::G4VPhysicalVolume,
+        tid: i32,
         pid: i32,
         energy: f64,
         total_deposit: f64,
@@ -408,7 +409,7 @@ impl<'a> RunAgent<'a> {
     ) {
         if let Some(deposits) = self.deposits.as_mut() {
             deposits.push(
-                volume, self.index - 1, pid, energy, total_deposit, point_deposit, start, end,
+                volume, self.index - 1, tid, pid, energy, total_deposit, point_deposit, start, end,
                 self.weight, &self.random_index
             )
         }
