+++ src/simulation/sampler.rs
@@ -205,6 +205,7 @@ impl Deposits {
         &mut self,
         volume: *const ffi::G4VPhysicalVolume,
         event: usize,
+        tid: i32,
         pid: i32,
         energy: f64,
         total_deposit: f64,
@@ -216,12 +217,13 @@ impl Deposits {
     ) {
         self.values.entry(volume)
             .and_modify(|e| e.push(
-                event, pid, energy, total_deposit, point_deposit, start, end, weight, random_index
+                event, tid, pid, energy, total_deposit, point_deposit, start, end, weight,
+                random_index
             ))
             .or_insert_with(|| {
                 let mut cell = DepositsCell::new(self.mode);
                 cell.push(
-                    event, pid, energy, total_deposit, point_deposit, start, end, weight,
+                    event, tid, pid, energy, total_deposit, point_deposit, start, end, weight,
                     random_index
                 );
                 cell
@@ -287,6 +289,7 @@ impl DepositsCell {
     fn push(
         &mut self,
         event: usize,
+        tid: i32,
         pid: i32,
         energy: f64,
         total_deposit: f64,
@@ -309,13 +312,14 @@ impl DepositsCell {
                 let random_index = *random_index;
                 if line_deposit > 0.0 {
                     let deposit = LineDeposit {
-                        event, pid, energy, start, end, value: line_deposit, weight, random_index
+                        event, tid, pid, energy, start, end, value: line_deposit, weight,
+                        random_index
                     };
                     deposits.line.push(deposit);
                 }
                 if point_deposit > 0.0 {
                     let deposit = PointDeposit {
-                        event, pid, energy, position: end, value: point_deposit, weight,
+                        event, tid, pid, energy, position: end, value: point_deposit, weight,
                         random_index
                     };
                     deposits.point.push(deposit);
@@ -336,6 +340,7 @@ impl DepositsCell {
 #[repr(C)]
 pub struct LineDeposit {
     event: usize,
+    tid: i32,
     pid: i32,
     energy: f64,
     value: f64,
@@ -349,6 +354,7 @@ pub struct LineDeposit {
 #[repr(C)]
 pub struct PointDeposit {
     event: usize,
+    tid: i32,
     pid: i32,
     energy: f64,
     value: f64,
