### 2025-06-12 59a53f6
Bump version: v1.1.2
