+++ src/python/calzone/__init__.py
@@ -2,4 +2,4 @@ from .calzone import *
 from .calzone import __doc__
 del calzone
 
-VERSION = "1.1.1"
+VERSION = "1.1.2"
