+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "calzone"
-version = "1.1.1"
+version = "1.1.2"
 edition = "2021"
 
 [lib]
