+++ .bumpversion.cfg
@@ -1,7 +1,7 @@
 [bumpversion]
 commit = True
 tag = False
-current_version = 1.1.1
+current_version = 1.1.2
 message = Bump version: v{new_version}
 
 [bumpversion:file:Cargo.toml]
