### 2025-06-25 #8485 @editorialbot
[REVIEW]: Calzone: A Geant4 Python wrapper for the simulation of outdoor particle detectors
**Submitting author:** @niess (<a href="http://orcid.org/0000-0001-7148-6819">Valentin Niess</a>)
**Repository:** https://github.com/niess/calzone
**Branch with paper.md** (empty if default branch): paper
**Version:** v1.1.4
