### 2025-06-25 #8485 @ekourlit
@niess I would need to ask you to reduce the length of the paper (it is currently 70% longer than our limit of 1000 words). I would propose to omit details on the section "Selected Calzone features".
