### 2025-07-02 43324a4
Reduce paper length
