+++ paper/paper.md
@@ -104,7 +104,7 @@ exemplified by, but not limited to, [Gamos][GAMOS] [@Arce:2014], [Gate][GATE]
 
 In the context of geosciences, we encountered specific issues that were not
 addressed by [Geant4][GEANT4], and only partially addressed by some of its
-derivatives. These issues, which motivated the development of
+derivatives. Some of these issues, which motivated the development of
 [Calzone][Calzone], are discussed hereafter.
 
 
@@ -119,44 +119,33 @@ The precision of MCPT computations is contingent upon the accuracy of the
 geometry description. In the context of geosciences, the aforementioned geometry
 includes the particle detector, which is depicted in a mechanical diagram (using
 a [CAD][CAD] software), as well as the study site, which is usually represented
-by a Digital Elevation Model ([DEM][DEM]). However, these data are not
-understood by [Geant4][GEANT4] by default, so they need to be transcribed. A
-generic method for transcribing composite data into [Geant4][GEANT4] geometry
-primitives is to delineate volumes of the same material (terrain, sensor,
-mechanical support, etc.) using surfaces approximated by triangular meshes. For
-instance, the [FreeCAD][FREECAD] software is able to export the detector parts
-as [STL][STL] files, which could then be re-read and transcribed into
-[`G4TessellatedSolids`][G4TS] using the [CADMesh][CADMESH] interface layer
-[@Poole:2012]. [Calzone][CALZONE] streamlines this process by defining a
-geometry format that serves as an intermediary. This format uses standard
-objects, including `dict`, `float`, `list`, and `str`, and integrates various
-mesh formats, such as [OBJ][OBJ], [STL][STL], [GeoTiff][GEOTIFF] and
-[Turtle][TURTLE]/[PNG][PNG] [@Niess:2020]. [Calzone][CALZONE] then translates
-this data into [Geant4][GEANT4] objects. During this process, [Calzone][CALZONE]
-takes care of sharing mesh data instead of copying it if the same mesh is used
-more than once.
+by a Digital Elevation Model ([DEM][DEM]). These data are not natively
+understood by [Geant4][GEANT4], requiring transcription. A generic approach is
+to delineate volumes of the same material (terrain, sensor, mechanical support,
+etc.) using surfaces approximated by triangular meshes. For instance, the
+[FreeCAD][FREECAD] software is able to export the detector parts as [STL][STL]
+files, which could then be re-read and transcribed into
+[`G4TessellatedSolids`][G4TS] (e.g., using [CADMesh][CADMESH] [@Poole:2012]).
+[Calzone][CALZONE] streamlines this process by defining a geometry format that
+serves as an intermediary. This format uses standard objects, including `dict`,
+`float`, `list`, and `str`, and integrates various mesh formats, such as
+[OBJ][OBJ], [STL][STL], [GeoTiff][GEOTIFF] and [Turtle][TURTLE]/[PNG][PNG]
+[@Niess:2020]. [Calzone][CALZONE] then translates this data into
+[Geant4][GEANT4] objects.
 
 ## Mesh specialisation
 
 The process of meshing a [DEM][DEM] with triangular facets introduces specific
-issues. To clarify this point, it should be noted that the Monte Carlo traversal
-of mesh structures requires optimisations in order to be a viable option. The
-[Geant4][GEANT4] software uses a [voxelisation][VOXEL] algorithm that
-subdivides the mesh volume into smaller units. This algorithm is effective for
-particles travelling short distances in terms of mesh extension, which is the
[…]
