### 2025-07-07 niess/calzone#1 @jbeirer
[REVIEW] JOSS Review of Calzone openjournals/joss-reviews#8485
Hi @niess,

many thanks for the submission — I had a chance to go through the paper and the repository, and overall I think this is a really solid contribution. Below are some comments based on the JOSS review criteria.

## Paper

The paper is clear and well-written. It does a good job explaining what Calzone is and why it’s useful. The summary is accessible to a broad audience, and the statement of need is convincing — the steep learning curve of Geant4 is a real barrier, and Calzone offers a practical way to lower it, especially for educational or small-scale detector projects.

Some minor comments and suggestions:

- L14: The name “Calzone” isn’t explained in the paper. Since it stands for “CALorimeter ZONE,” it might be helpful to include that in parentheses the first time the name is mentioned, to make the origin and purpose clearer to readers.
- L25: It could be helpful to clarify the units used for energy, position, etc., in the code example (e.g., MeV, mm?) to avoid any ambiguity for users unfamiliar with Geant4 defaults.
- L66: I suggest removing "the" or replacing "the like" with "such as" for correct grammar.
- L80: I suggest changing "modelling of particles sources" to "modelling of particle sources" for correct grammar.
- L94: "This is contribution no. XXX” seems to be a placeholder — I assume this will be updated in later revisions

## Functionality

Everything worked well when I tested it. Installation via pip is straightforward, and the usage examples are clear and easy to follow. Currently, only [manylinux wheels](https://pypi.org/project/calzone/#files) are available on PyPI, which means installation via pip fails on macOS. It would be great to also provide wheels for both Intel and ARM-based macOS systems to support a broader range of users.

## Tests

I see there are a number of well written tests [in the repository ](https://github.com/niess/calzone/tree/master/tests). However, there’s currently no continuous integration (CI) system set up to automatically run these tests.

Since JOSS strongly encourages an automated test suite with CI (e.g., GitHub Actions), I’d recommend adding a simple workflow to ensure test coverage is automatically monitored. It would also be useful to include the example scripts as part of the CI checks, to ensure they continue to work as expected. This would make it easier to catch regressions and help contributors verify that changes don’t break core functionality.

## Community guidelines

Please add community guidelines (e.g., a CONTRIBUTING.md file) to clarify how users can contribute, report issues, or seek support.

## Documentation

The documentation is generally good, with clear explanations and good API coverage. As a small improvement, the authors could consider adding a quick start section to the README to help new users get up and running more quickly.

Overall, this is a strong submission, and I would **recommend acceptance after minor revisions**.

Best,
Joshua
