+++ pytest.ini
@@ -1,5 +1,7 @@
 [pytest]
 addopts = -l
+markers =
+    requires_data
 testpaths =
     tests
     docs/source
