+++ pyproject.toml
@@ -45,13 +45,25 @@ test-requires = [ "Pillow", "pytest" ]
 before-all = """
 curl -sSf https://sh.rustup.rs -o rustup.sh && \
 sh rustup.sh -y && \
-curl -sSf -L https://github.com/niess/calzone/releases/download/geant4/geant4-11.2.2-manylinux2014-x86_64.tgz -o geant4.tgz && \
-tar -xzf geant4.tgz\
+curl -sSf -L https://github.com/niess/calzone/releases/download/geant4/geant4-11.3.0-manylinux2014_$(uname -m).tar.gz -o geant4.tar.gz && \
+tar -xzf geant4.tar.gz\
 """
 repair-wheel-command = """
-LD_LIBRARY_PATH=$PWD/geant4/lib64/:$LD_LIBRARY_PATH \
+LD_LIBRARY_PATH=$PWD/geant4/lib64/:$PWD/geant4/lib/:$LD_LIBRARY_PATH \
 auditwheel repair -w {dest_dir} {wheel}\
 """
-build = "cp37-manylinux_x86_64"
+build = [ "cp37-manylinux_x86_64", "cp37-manylinux_aarch64" ]
 environment = "PATH=$HOME/.cargo/bin:$PATH"
 manylinux-x86_64-image = "manylinux2014"
+
+[tool.cibuildwheel.macos]
+before-all = """
+curl -sSf -L https://github.com/niess/calzone/releases/download/geant4/geant4-11.3.0-macosx_11_0_$(uname -m).tar.gz -o geant4.tar.gz && \
+tar -xzf geant4.tar.gz\
+"""
+build = "cp38-macosx_*"
+repair-wheel-command = """
+DYLD_LIBRARY_PATH=$PWD/geant4/lib64/:$PWD/geant4/lib/:$DYLD_LIBRARY_PATH \
+delocate-wheel --require-archs {delocate_archs} -w {dest_dir} -v {wheel}
+"""
+environment = { MACOSX_DEPLOYMENT_TARGET = "11.0" }
