+++ .github/workflows/package.yml
@@ -22,6 +22,7 @@ jobs:
   build_sdist:
     name: Build sdist
     runs-on: ubuntu-latest
+
     if: ${{ inputs.sdist }}
     steps:
       - uses: actions/checkout@v4
@@ -41,7 +42,8 @@ jobs:
     runs-on: ${{ matrix.os }}
     strategy:
       matrix:
-        os: [ubuntu-22.04]
+        os: [ubuntu-latest, ubuntu-24.04-arm, macos-13, macos-latest]
+
     if: ${{ inputs.wheel }}
     steps:
       - uses: actions/checkout@v4
@@ -49,7 +51,7 @@ jobs:
           submodules: recursive
 
       - name: Build wheel
-        uses: pypa/cibuildwheel@v2.16
+        uses: pypa/cibuildwheel@v2.23.3  # 2.23.3: supports Python 3.7.
 
       - uses: actions/upload-artifact@v4
         with:
@@ -63,6 +65,7 @@ jobs:
     environment: pypi
     permissions:
       id-token: write
+
     if: ${{ inputs.upload }}
     steps:
       - uses: actions/download-artifact@v4
