### 2025-07-24 bcbc73f
Add examples to tests
