+++ pytest.ini
@@ -1,7 +1,10 @@
 [pytest]
 addopts = -l
 markers =
+    example
     requires_data
+    requires_display
+    requires_goupil
 testpaths =
     tests
     docs/source
