+++ pyproject.toml
@@ -35,11 +35,7 @@ target = "calzone.calzone"
 
 # Build options for Python wheels.
 [tool.cibuildwheel]
-test-command = [
-  'pytest --doctest-glob="*.rst" -v -m "not requires_data" {project}/tests {project}/docs/source',
-  'python {project}/examples/geometry/topography/generate.py',
-  'test -f "{project}/examples/geometry/topography/meshes/terrain.png"',
-]
+test-command = 'pytest --doctest-glob="*.rst" -v -m "not requires_data" {project}/tests {project}/docs/source'
 test-requires = [ "Pillow", "pytest" ]
 
 [tool.cibuildwheel.linux]
@@ -65,7 +61,7 @@ tar -xzf geant4.tar.gz\
 build = "cp38-macosx_*"
 repair-wheel-command = """
 DYLD_LIBRARY_PATH=$PWD/geant4/lib64/:$PWD/geant4/lib/:$DYLD_LIBRARY_PATH \
-delocate-wheel --require-archs {delocate_archs} -w {dest_dir} -v {wheel}
+delocate-wheel --require-archs {delocate_archs} -w {dest_dir} -v {wheel}\
 """
 environment = { MACOSX_DEPLOYMENT_TARGET = "11.0" }
 
