### 2025-07-24 786667e
Automate Geant4 download
