### 2025-07-25 2d37274
Use $GEANT4_PREFIX
