### 2025-07-28 b6fede0
Parse datasets from geant4.sh script
