+++ Cargo.toml
@@ -22,7 +22,7 @@ process_path = "0.1"
 pyo3 = { version = "0.21", features = ["abi3", "extension-module"] }
 rand = "0.8"
 rand_pcg = "0.3"
-regex = "1.10"
+regex.workspace = true
 rmp-serde = "1.3"
 reqwest.workspace = true
 serde = { version = "1.0", features = [ "derive"] }
@@ -32,12 +32,14 @@ temp-dir.workspace = true
 [build-dependencies]
 cxx-build = "1.0"
 flate2.workspace = true
+regex.workspace = true
 reqwest.workspace = true
 tar.workspace = true
 temp-dir.workspace = true
 
 [workspace.dependencies]
 flate2 = "1.0"
+regex = "1.10"
 reqwest = { version = "0.12", features = ["blocking", "rustls-tls"], default-features = false }
 tar = "0.4"
 temp-dir = "0.1"
