### 2025-07-28 154c61b
Close display when testing
