+++ tests/test_examples.py
@@ -1,16 +1,27 @@
+import os
 from pathlib import Path
 import pytest
+import shutil
 import subprocess
 import sys
+import tempfile
 
 
 PREFIX = Path(__file__).parent.parent
 
 
-def run(path):
+def run(path, prepend=None):
     """Run example script."""
 
     path = PREFIX / f"examples/{path}"
+    if prepend is not None:
+        tmp = tempfile.TemporaryDirectory()
+        shutil.copytree(path.parent, tmp.name, dirs_exist_ok=True)
+
+        path = Path(tmp.name) / path.name
+        with path.open("a") as f:
+            f.write(f"{os.linesep}{prepend}")
+
     command = f"{sys.executable} {path}"
     r = subprocess.run(command, shell=True, capture_output=True)
     if r.returncode != 0:
@@ -54,7 +65,10 @@ def test_topography():
 def test_trajectograph_muons():
     """Test the trajectograph muons example."""
 
-    run("muon/trajectograph/run.py")
+    run("muon/trajectograph/run.py", prepend="""
+import calzone_display
+calzone_display.close()
+""")
 
 @pytest.mark.example
 @pytest.mark.requires_data
