### 2025-07-28 ffcb122
Update the installation instructions
