### 2025-07-29 93e1641
Test goupil example
