+++ CONTRIBUTING.md
@@ -1,0 +1,21 @@
+# Contributing to Calzone
+
+Calzone is an [open-source][OPEN_SOURCE] project, which means that you, as a
+user, are welcome to contribute by reporting any [issues][ISSUES] or suggesting
+[new features][ISSUES].
+
+The Calzone [source code](src) is written in [Rust][RUST] with a pinch of C++
+(for interfacing to Geant4, using the [cxx][CXX] crate). The Python user
+interface is generated using the [PyO3][PYO3] crate. If you have knowledge of
+[Rust][RUST] (and C++, Python, and [Geant4][GEANT4]), or wish to start learning
+[Rust][RUST], submitting fixes or new features as [pull requests][PR] would be
+even more welcomed.
+
+
+[CXX]: https://github.com/dtolnay/cxx
+[GEANT4]: https://geant4.web.cern.ch/
+[ISSUES]: https://github.com/niess/calzone/issues
+[OPEN_SOURCE]: https://en.wikipedia.org/wiki/Open_source
+[PR]: https://github.com/niess/calzone/pulls
+[PYO3]: https://github.com/PyO3/pyo3
+[RUST]: https://www.rust-lang.org
