+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "calzone"
-version = "1.1.2"
+version = "1.1.3"
 edition = "2021"
 
 [lib]
