+++ pyproject.toml
@@ -15,7 +15,7 @@ classifiers = [
     "Intended Audience :: Science/Research",
     "Topic :: Scientific/Engineering :: Physics"
 ]
-version = "1.1.2"
+version = "1.1.3"
 requires-python = ">=3.7.0"
 dependencies = [
     "numpy >= 1.6.0",
