### 2025-08-01 niess/calzone#1 @niess
Hello Joshua (@jbeirer),

Thank you for confirming that Calzone runs on macOS.

Concerning the test suite, I am reluctant to trigger it "blindly", because running these workflows does have some cost, although this isn't explicit since GitHub doesn't charge us for it. For instance, the Package workflow takes ~1 h (cumulated over 6 platforms), most of which is for compiling Calzone's dependencies.

This could be lowered by testing a single platform, and maybe by caching the "build" directory. But, that's what I am already doing locally, while I am developping.

Best,

Valentin
