### 2025-09-09 #8485 @ekourlit
Hi @jbeirer and @peremato , I see there are still outstanding items on you checklists. Would you like to request anything further from the author?
Currently, the only open issue, that we would need to close to proceed is: https://github.com/niess/calzone/issues/1.
