### 2025-09-09 #8485 @ekourlit
That's great! Could you @jbeirer (or @niess) then close the https://github.com/niess/calzone/issues/1 issue please?
