### 2025-09-09 niess/calzone#2 @peremato
[REVIEW] JOSS Review of Calzone (openjournals/joss-reviews#8485)
First at all apologies for the delay due to summer holidays and change of employment status.

Very nice work. I have installed the software and run some of the examples. It is impressive how easy is to develop and run a simulation using this Python package compared with the original Geant4 C++ interface. It is certainly a very solid contribution.

Following there are some comments based on the JOSS criteria and some suggestions:

### Paper

The paper is clear and well-written. It summarizes well the motivation and purpose of Calzone and its basic functionality, as well as its place with respect other existing packages on top of Geant4.The list of references is complete and they cite everything appropriately.

Some minor comments and suggestions:

- L34: Perhaps the link for **validation** should be pointing to https://geant4.web.cern.ch/collaboration/working_groups/physval-wg
- L89. I do not understand very well the purpose of the section **Author contributions** specifying in detail the contribution of each author. Perhaps it is common in JOSS papers, but it is unusual in other papers.

### Tests

There are a number of tests available. I would have liked to see in the README how to run them. I know that is trivial but for somebody not used to it would certainly help.

Some comments:

- Running all tests successfully requires to have `calzone-display` and `goupil` installed. I wonder whether it can be detected automatically.
```
FAILED tests/test_examples.py::test_goupil - RuntimeError: Traceback (most recent call last):
FAILED tests/test_examples.py::test_trajectograph_muons - RuntimeError: Traceback (most recent call last):
```
- Same comment as the other reviewer concerning the CI

### Examples
I have managed to run some of the examples on my laptop (MacOS M2). However I failed to run the display.
```
% calzone display detector.toml

thread '<unnamed>' panicked at /Users/runner/.cargo/registry/src/index.crates.io-1949cf8c6b5b557f/winit-0.30.12/src/platform_impl/macos/event_loop.rs:221:14:
on macOS, `EventLoop` must be created on the main thread!
stack backtrace:
   0: __rustc::rust_begin_unwind
   1: core::panicking::panic_fmt
   2: core::option::expect_failed
   3: winit::platform_impl::macos::event_loop::EventLoop<T>::new
   4: <bevy_winit::WinitPlugin<T> as bevy_app::plugin::Plugin>::build
   5: bevy_app::app::App::add_boxed_plugin
   6: bevy_app::plugin_group::PluginGroupBuilder::finish
   7: bevy_app::app::App::add_plugins
note: Some details are omitted, run with `RUST_BACKTRACE=full` for a verbose backtrace.

thread '<unnamed>' panicked at src/app.rs:103:23:
called `Result::unwrap()` on an `Err` value: Any { .. }
stack backtrace:
   0: __rustc::rust_begin_unwind
   1: core::panicking::panic_fmt
   2: core::result::unwrap_failed
   3: calzone_display::app::<impl calzone_display::app::stop::MakeDef>::_PYO3_DEF::trampoline
   4: _cfunction_vectorcall_NOARGS
   5: _atexit_callfuncs
   6: __Py_Finalize
   7: _Py_Exit
   8: __PyErr_PrintEx
   9: __PyRun_SimpleFileObject
  10: __PyRun_AnyFileObject
  11: _pymain_run_file_obj
  12: _pymain_run_file
  13: _Py_RunMain
  14: _pymain_main
  15: _Py_BytesMain
note: Some details are omitted, run with `RUST_BACKTRACE=full` for a verbose backtrace.
Exception ignored in atexit callback <built-in function stop>:
pyo3_runtime.PanicException: called `Result::unwrap()` on an `Err` value: Any { .. }
```

### Contributing

CONTRIBUTING.md file exists with clear messages

### Documentation

The documentation is good and the API is complete as far I can see. The tables to specify the geometry is a bit convoluted and complex. A user guide oriented section with one or two examples would be nice to have. In addition, perhaps in the README, you can mention the existence of the examples, their purpose and how to run them.

### Conclusion
In summary it is a strong submission. I would recommend for acceptance after minor revisions.
