### 2025-09-19 7516f0a
Extend geometry docs
