+++ docs/source/requirements.txt
@@ -2,3 +2,4 @@ calzone
 numpy
 sphinx
 sphinx-book-theme
+sphinx-subfigure
