+++ docs/source/geometry.rst
@@ -9,11 +9,172 @@ A `Geant4`_ Monte Carlo geometry consists of a hierarchy of nested
 and :external:py:class:`str`) that have associated representations in common
 configuration languages, such as `JSON`_, `TOML`_ or `YAML`_.
 
-A comprehensive account of the Calzone geometry format is presented in the
-following sections, accompanied by `TOML`_ code examples. For a more pragmatic
-approach, one might alternatively consult the :bash:`examples/` folder, which is
-distributed with the Calzone `source <Examples_>`_.
+In the following section, we illustrate Calzone's geometry representation with
+some of the examples distributed with the `source <Examples_>`_. For a more
+comprehensive account of the Calzone geometry definition, please refer to the
+`Reference`_ section hereafter.
 
+.. note::
+
+   While Calzone supports various geometry formats, the examples provided in
+   this section are all in `TOML`_ format.
+
+----
+
+Examples
+********
+
+Flat air-soil interface
+-----------------------
+
+To begin with a basic example, let us consider a flat air-soil interface (as in
+the `gamma/benchmark <ExampleBenchmark_>`_ example). This is represented with
+Calzone using a :bash:`geometry.toml` file, with the following content
+
+.. literalinclude:: include/benchmark-geometry.toml
+   :language: toml
+   :lines: 1-8
+
+The root volume is defined by an axis-aligned cubic box of 2 km extension (i.e.
+:python:`2E+05` cm), named :toml:`Environment` and filled with :toml:`"G4_AIR"`
+(from the Geant4 `NIST`_ database). The root volume (:toml:`Environment`)
+contains a daughter volume (:toml:`Ground`), which is also of box shape, with a
+square base of 2 km sides and a height of 1 km. The origin (i.e. the box centre)
+of the :python:`Ground` volume is offset by 500 m along the z-axis w.r.t. its
+mother volume (as specified by the :toml:`position` attribute). As a result, the
+:toml:`Ground` volume occupies the lower half of the root volume.
+
+.. tip::
+
+   Calzone supports various volume shapes, not only boxes. Please refer to the
+   `Shape definition`_ section below for a description of the available volume
+   shapes.
+
+Following the same box logic, a second  :toml:`Detector` volume (:math:`20
+\times 20 \times 10 \, \mathrm{m}^3`) is placed within the root volume, with its
+base positioned 5 cm above the top of the ground volume. This is achieved as
+
+.. literalinclude:: include/benchmark-geometry.toml
+   :language: toml
+   :lines: 9-
+
+The resulting geometry is represented in :numref:`fig-benchmark-geometry` below.
+
+.. warning::
+
+   It is the user's responsibility to ensure that daughter volumes do not
+   collide with each other or with their mother volume. This can be tested by
+   running the :py:meth:`Geometry.check <calzone.Geometry.check>` method.
+
+.. subfigure:: AB
+   :gap: 8px
+   :layout-sm: A|B
+   :name: fig-benchmark-geometry
+   :subcaptions: above
+
+   .. image:: include/benchmark-geometry.svg
+      :alt: (A)
+
+   .. image:: include/benchmark-geometry.png
+      :alt: (B)
+
+   Representations of the `gamma/benchmark <ExampleBenchmark_>`_ example. (A)
+   Cross-section along the (xOz) plane. (B) Calzone display view.
+
+
+Including a topography
+----------------------
+
+Let us now consider a more realistic example, in which the air-soil interface is
+described by a Digital Elevation Model (`DEM`_), as in the `geometry/topography
[…]
