+++ docs/source/include/trajectograph-geometry.toml
@@ -1,0 +1,1 @@
+./../../../examples/muon/trajectograph/geometry.toml
