+++ docs/source/include/trajectograph-layer.toml
@@ -1,0 +1,1 @@
+../../../examples/muon/trajectograph/layer.toml
