+++ docs/source/include/benchmark-geometry.toml
@@ -1,0 +1,1 @@
+./../../../examples/gamma/benchmark/geometry.toml
