### 2025-10-06 95ac374
Improve bounding box
