### 2025-10-06 d10cfde
Add shapes examples
