+++ .bumpversion.toml
@@ -1,0 +1,32 @@
+[tool.bumpversion]
+commit = true
+current_version = "1.1.3"
+message = 'Bump version: v{new_version}'
+
+[[tool.bumpversion.files]]
+filename = "Cargo.toml"
+search = '''
+[package]
+name = "calzone"
+version = "{current_version}"
+'''
+replace = '''
+[package]
+name = "calzone"
+version = "{new_version}"
+'''
+
+[[tool.bumpversion.files]]
+filename = 'pyproject.toml'
+search = 'version = "{current_version}"'
+replace = 'version = "{new_version}"'
+
+[[tool.bumpversion.files]]
+filename = 'src/python/calzone/__init__.py'
+search = 'VERSION = "{current_version}"'
+replace = 'VERSION = "{new_version}"'
+
+[[tool.bumpversion.files]]
+filename = 'docs/source/conf.py'
+search = 'release = "{current_version}"'
+replace = 'release = "{new_version}"'
