+++ .bumpversion.cfg
@@ -1,17 +1,0 @@
-[bumpversion]
-commit = True
-tag = False
-current_version = 1.1.3
-message = Bump version: v{new_version}
-
-[bumpversion:file:Cargo.toml]
-parse = "version = \"(?P<major>\d+)\.(?P<minor>\d+)\.(?P<patch>\d+)\""
-
-[bumpversion:file:docs/source/conf.py]
-parse = "release = '(?P<major>\d+)\.(?P<minor>\d+)\.(?P<patch>\d+)'"
-
-[bumpversion:file:pyproject.toml]
-parse = "version = (?P<major>\d+)\.(?P<minor>\d+)\.(?P<patch>\d+)"
-
-[bumpversion:file:src/python/calzone/__init__.py]
-parse = "VERSION = (?P<major>\d+)\.(?P<minor>\d+)\.(?P<patch>\d+)"
