### 2025-10-07 71dd539
Rename core module
