+++ Cargo.toml
@@ -5,6 +5,7 @@ edition = "2021"
 
 [lib]
 crate-type = ["cdylib"]
+name = "_core"
 
 [dependencies]
 bvh = "0.9"
