+++ src/python/calzone/__init__.py
@@ -11,4 +11,4 @@ def init():
 init()
 del init
 
-VERSION = "1.1.3"
+VERSION = "1.1.4"
