+++ .bumpversion.toml
@@ -1,6 +1,6 @@
 [tool.bumpversion]
 commit = true
-current_version = "1.1.3"
+current_version = "1.1.4"
 message = 'Bump version: v{new_version}'
 
 [[tool.bumpversion.files]]
