+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "calzone"
-version = "1.1.3"
+version = "1.1.4"
 edition = "2021"
 
 [lib]
