### 2025-10-08 7e287fc
Update display example
