### 2025-10-08 #8485 @niess
Hello, following Pere's review (niess/calzone#2), I slightly updated the paper (a hyperlink, and the example figure), as well as the project documentation, and the `display` extension which should now (hopefully) also run on macOS (it does on Linux & Windows).

Thank you all, for your time reviewing this work.
