### 2025-10-16 niess/calzone#2 @peremato
Thanks very much for the nice improvements. I tested the display in my MacOS with M2 and works nicely. The documentation on the display is very useful. Without it I couldn't guess the clicks/gestures to interact with the display since Bevy is new to me.
The user guide for the geometry in the documentation is also very useful. No way to guess that you needed to include `toml` files to replicate daughter volumes in a mother  volume.

I think everything is now completed. I am closing the issue.
