### 2025-10-29 #8485 @ekourlit
@niess the paper now passes the checks! However, since we have updates, I would recommend you cut a new release and update Zenodo accordingly.

Congratulations once again!
