+++ paper/paper.md
@@ -25,7 +25,7 @@ bibliography: paper.bib
 
 # Summary
 
-The transport of high-energy particles (e.g., [gamma-rays][GAMMA_RAY]) through
+The transport of high-energy particles (e.g., gamma-rays) through
 ordinary matter is an inherently stochastic process, with individual collisions
 described within the framework of quantum field theory. The resolution of
 such transport problems is facilitated by the use of Monte Carlo methods,
@@ -37,10 +37,10 @@ particles through matter.
 [Calzone][CALZONE] (CALorimeter ZONE) is a MCPT [Python][PYTHON] package built
 on top of [Geant4][GEANT4]. It was developed in the context of geosciences with
 the objective of studying the emission of radioactivity from volcanoes
-[@Terray:2020], and in particular to simulate the response of [gamma][GAMMA]
+[@Terray:2020], and in particular to simulate the response of gamma
 spectrometers deployed in the field. To this end, [Calzone][CALZONE] was
 developed in conjunction with [Goupil][GOUPIL] [@Niess:2024], a backward
-[gamma][GAMMA] transport engine, and is interoperable with the latter. Yet, both
+gamma transport engine, and is interoperable with the latter. Yet, both
 packages can be used entirely independently, if necessary.
 
 [Calzone's][CALZONE] interface has been designed with simplicity in mind. Source
@@ -63,10 +63,10 @@ deposits = simulation.run(particles).deposits
 ```
 
 [Calzone][CALZONE] encourages the use of meshes to describe the Monte Carlo
-geometry. Various mesh formats are supported, such as [OBJ][OBJ], [STL][STL],
-[GeoTIFF][GEOTIFF] and [Turtle][TURTLE]/[PNG][PNG] [@Niess:2020]. These formats
-can be used to encode the components of a detector (exported from a computer-aided design, or [CAD][CAD],
-scheme) or a digital elevation model ([DEM][DEM]) describing the surrounding
+geometry. Various mesh formats are supported, such as OBJ, STL,
+GeoTIFF and [Turtle][TURTLE]/PNG [@Niess:2020]. These formats
+can be used to encode the components of a detector (exported from a computer-aided design, or CAD,
+scheme) or a digital elevation model (DEM) describing the surrounding
 terrain. Additionally, [Calzone][CALZONE] features an interactive display
 ([`calzone-display`][CALZONE_DISPLAY]) that allows users to navigate through the
 Monte Carlo geometry and to inspect Monte Carlo tracks (see, e.g.,
@@ -87,7 +87,7 @@ or space sciences).
 
 ![
 Example of [`calzone-display`][CALZONE_DISPLAY]. The background image comprises
-a digital elevation model ([DEM][DEM]) of the [Masaya][MASAYA] volcano, derived
+a digital elevation model (DEM) of the Masaya volcano, derived
 from photogrammetry measurements. The grey box on the volcano ridge
 (bottom-left) corresponds to a gamma-spectrometer (located at
 $11.983056^\circ$N, $86.172815^\circ$W), the details of which
@@ -122,39 +122,39 @@ package, along with the specific issues that these features address.
 The precision of MCPT computations is contingent upon the accuracy of the
 geometry description. In the context of geosciences, the aforementioned geometry
 includes the particle detector, which is depicted in a mechanical diagram (using
[…]
