### 2025-11-04 #8485 @kyleniemeyer
@niess thank you. The version number also does not need to change unless the software itself has been changed (since the version only applies to the software).

No problem on keeping the paper on a separate, unarchived branch. JOSS archives the paper, and the archive+version number only refer to the reviewed software itself, not the full archive / all versions / all branches.
